+++ keybot/main_test.go
@@ -18,26 +18,6 @@ func TestAddCommands(t *testing.T) {
 	addCommands(bot)
 }
 
-func TestBuildPlease(t *testing.T) {
-	out, err := kingpinHandler([]string{"build", "please"})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if out != "Dry Run: Doing that would run `/bin/launchctl` with args: [start keybase.prerelease]" {
-		t.Errorf("Unexpected output: %s", out)
-	}
-}
-
-func TestPromoteRelease(t *testing.T) {
-	out, err := kingpinHandler([]string{"release", "promote", "1.2.3"})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if out != "Dry Run: Doing that would run `/bin/launchctl` with args: [start keybase.prerelease.promotearelease]" {
-		t.Errorf("Unexpected output: %s", out)
-	}
-}
-
 func TestInvalidUsage(t *testing.T) {
 	out, err := kingpinHandler([]string{"release", "oops"})
 	if err != nil {
